Extracted UserInterface.java
